requirements.txt
langchain
langchain-core
langchain-openai
langchain-chroma
python-dotenv
pandas